added search items method and organized the ebay buy
